feat: add router and the NotFound page
