+++ src/components/NotFound/NotFound.tsx
@@ -1,0 +1,17 @@
+import { Link } from 'react-router';
+import style from './NotFound.module.css';
+
+export default function NotFound() {
+  return (
+    <div className={style.page_wrapper}>
+      <div className={style.not_found_info}>
+        <div className={style.not_found}>Page not found</div>
+        <Link to="/">
+          <button type="button" className={style.to_main_button}>
+            Back to the main page
+          </button>
+        </Link>
+      </div>
+    </div>
+  );
+}
