+++ src/main.tsx
@@ -3,6 +3,7 @@ import { createRoot } from 'react-dom/client';
 import './index.css';
 import ErrorBoundary from './components/ErrorBoundary/ErrorBoundary.tsx';
 import { App } from './App.tsx';
+import { BrowserRouter } from 'react-router';
 
 const rootElement = document.getElementById('root');
 
@@ -10,7 +11,9 @@ if (rootElement instanceof HTMLElement)
   createRoot(rootElement).render(
     <StrictMode>
       <ErrorBoundary>
-        <App />
+        <BrowserRouter>
+          <App />
+        </BrowserRouter>
       </ErrorBoundary>
     </StrictMode>
   );
